feat: add project structure
features a layout based on clean architecture: commands for use cases, io for adapters and core for main functionality. Main function stablishes a common structure that simplifies adding commands on the future by enforcing some basic rules (check 'CliCommand' trait). Init command is provided currently as an example and test of this structure